+++ server/src/index.ts
@@ -49,14 +49,19 @@ app.post('/workouts', async (req, res, next) => {
 
         const date = getDayFromString(body.date);
 
+        const existing = await getWorkoutsFrom(date);
+        if (existing.length > 1) throw tooManyWorkoutsExistsWith(date);
+
+        const existingExercises = Object.keys(existing[0].exercises);
+        if (existingExercises.length - 2 >= body.exercises.length) {
+            throw new Error('you are removing too many exercises');
+        }
+
         if (body.exercises.length == 0) {
             await deleteWorkout(date);
             res.status(200).json({ 'message': 'success' });
             return;
         }
-
-        const existing = await getWorkoutsFrom(date);
-        if (existing.length > 1) throw tooManyWorkoutsExistsWith(date);
 
         const isUpdating = existing.length == 1;
